Added more configs for additional mlms, and implementations for topk and batchtopk activations; they don't seem to work that well in practice though, so I'll probably just stick with gated sae for whatever analysis we end up doing. TODO: maybe implement matryoshka sae and see if those perform better?

+++ sae/backbone.py
@@ -41,7 +41,7 @@ class MLLMBackbone:
         # -------------------------
         # Extract hidden states from specified layer_idx
         # -------------------------
-        x = hidden_states[layer_idx]  # (B, T, H)
+        x = hidden_states[layer_idx].float()  # (B, T, H)
 
         # -------------------------
         # Mask out special tokens / attention mask
